requirements.txt
google-adk==1.14.0
google-cloud-aiplatform
python-dotenv
requestspsycopg2-binary
